kernels: four more density shards, and drop the browser lp__ warnings
Splitting again, since the instantiations are the product and splitting is
the only lever on compile time. densities_common.cpp held both halves of
the tier-3 list and every hand-written kernel; the GLMs and the
integer-outcome densities now have their own translation units, and the
scalar lists split in two each. Nine shards, and the largest object goes
from 8.8 MB to 5.2 MB.

Balance them by argument count, not by entry count: measured, a four
argument tier-3 density costs about twice a three-argument one, and the
first split put four of the four-argument ones in the same file (6.2 MB
against 1.7 MB for its sibling). Moving two across evened it out.

The browser no longer reports a shifted lp__, so the notes saying it does
are gone: the footer on the demo page, the paragraph in js/README.md, and
the README section. fit.exactLp and stanli_exact_lp() stay, because
STANLI_LITE_LP is still available; they now report true.

Browser numbers re-measured for the exact build: 5.79 MB raw, 1.52 MB
gzipped, of which the core runtime is 2.26 MB raw and 0.69 MB gzipped and
the densities are the other 55% of the compressed download. The README's
per-library attribution table for the wasm binary is replaced by that
split, because the shipped wasm is stripped and the table could not be
regenerated honestly.
